Fix cursor disappearance in QLineEdit on Mac when deleting all text
On Mac OS X, if all text in the QLineEdit was selected and then deleted, cursor
visibility was not updated, and so the cursor remained hidden. Fixed
to update cursor visibility also when the text is empty.

Task-number: QTBUG-13169
Change-Id: Id52a20b07bb96609a78c42eb630ee2b20ed7cbcb
Reviewed-by: Jiang Jiang <jiang.jiang@nokia.com>
(cherry picked from commit 181456d0a31b7250da97eafba75e6bc657391777)