Fixed qtmodule-configtests on Windows.
open() to "|-" and "-|" are unsupported on Windows (see `perldoc
perlport'), so don't do that.

Let the tests write a 'config.log' file which contains the output
of the test commands.

Brings back the part fixing the pipe logic from commit
f865dc1ae44c88f6965acd09bafb01829c35447b which was reverted.

Change-Id: I5060a0885702d925001b98f2d4e84743d6ff226e
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>