Allow generic plugins to set defaults for window system properties
In order for generic plugins to set defaults on "startup" time - such
as the meego integration plugin to set the correct screen orientation -
it is necessary to construct the plugins when the application startup is
done. Then the plugin can "inject" the values the usual way, using
QWindowSystemInterface::handle*Change. Afterwards we need to process those
events - take them from the window system event queue and let QGuiApplication
process them.

Change-Id: I84de022ad565a33ae3ef5dfc34f540d6bf488b03
Reviewed-by: Samuel Rødal <samuel.rodal@nokia.com>